fastapi_framework/requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
loguru==0.6.0
tortoise-orm[aiomysql]==0.18.1
aerich==0.6.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5